+++ src/components/form/checkbox/checkbox.vue
@@ -1,7 +1,7 @@
 <template>
   <div>
     <label class="ods-checkbox" :class="{ 'ods-checkbox--error': errorMessage }">
-      <input type="checkbox" v-model="checkedModel" :value="value" :name="name" :aria-describedby="conditionalErrorId" />
+      <input type="checkbox" v-model="checkedModel" :value="value" :name="name" :aria-describedby="conditionalErrorId" :aria-required="ariaReq ? 'true' : null" />
       <span class="ods-checkbox__checkmark"></span>
       <span class="ods-checkbox__text">{{ label }}</span>
     </label>
@@ -36,6 +36,9 @@ export default {
     errorId: {
       type: String,
       required: true,
+    },
+    ariaReq: {
+      type: Boolean,
     },
   },