+++ src/components/form/textarea/textarea.vue
@@ -2,7 +2,7 @@
   <div class="ods-textarea" :class="{ 'ods-textarea--error': errorMessage }">
     <label class="ods-textarea__label">
       {{ label }}
-      <textarea v-model="textModel" :name="name" :aria-describedby="conditionalErrorId" class="ods-textarea__textarea" :rows="rows" :placeholder="placeholder"></textarea>
+      <textarea v-model="textModel" :name="name" :aria-describedby="conditionalErrorId" :aria-required="ariaReq ? 'true' : null" class="ods-textarea__textarea" :rows="rows" :placeholder="placeholder"></textarea>
     </label>
     <div v-if="errorMessage" class="ods-input__error-message" :id="errorId">{{ errorMessage }}</div>
   </div>
@@ -39,6 +39,9 @@ export default {
     errorId: {
       type: String,
       required: true,
+    },
+    ariaReq: {
+      type: Boolean,
     },
   },